+++ src/layouts/ReceivedDetail/index.tsx
@@ -3,13 +3,13 @@
 import { FaDownload, FaPrint } from 'react-icons/fa6'
 import Card from '../../components/Card'
 import { StrapiRecord } from '../../utils/strapi'
-import { CardAttributes } from '../../utils/strapi/card/server'
 import { mediaRecordsToUrlSet } from '../../utils/strapi/strapiImage'
 import * as styles from './index.css'
 import Link from 'next/link'
 import Renderer from '../../components/Card/Renderer'
 import { useState } from 'react'
 import Print from '../../components/Print'
+import { CardAttributes } from '../../utils/strapi/card'
 
 type ReceivedDetailProps = {
   cardRecord: StrapiRecord<Omit<CardAttributes, 'creator'>>
